parse user id from input string to long

+++ my-app-be-application/src/main/java/xyz/shiqihao/app/service/UserService.java
@@ -15,7 +15,7 @@ import xyz.shiqihao.app.model.User;
 public class UserService {
     private UserDao userDao;
 
-    public UserDto findById(int id) {
+    public UserDto findById(long id) {
         User user = userDao.findById(id);
         if (user == null) {
             return null;
